Added SJ modes
+ Added Endless
+ Added Random difficulties
+ Added Classic
+ Added Points/Paragons
+ Added Multipliers
* Changed SJ Menu

+++ src/main/java/com/spleefleague/parkour/game/GameHistory.java
@@ -42,7 +42,7 @@ public class GameHistory extends DBEntity implements DBSaveable {
         for (ParkourPlayer sjp : battle.getPlayers()) {
             players[i++] = new PlayerData(sjp.getUniqueId(), battle.getData(sjp).getFalls(), sjp == winner, !activePlayers.contains(sjp));
         }
-        this.duration = battle.getDuration();
+        this.duration = (int)battle.getDuration();
         startDate = new Date(System.currentTimeMillis() - this.duration * 50);
         this.arena = battle.getArena().getName();
         this.superjumpMode = battle.getArena().getParkourMode();
